mover strcmp_ci para utils.h, incluindo utils.h em verificador.c e hash.c, porque evita duplicação de código.
